Windows password analyser (#872)
* Shadow cracking worker

* Not finding a shadow file is still a successful task run

* Windows password cracker

* Match Linux task

* Abort was removed

* Include John on path for checking dependencies

* Move Windows analysis to work on Partions

* Address some of the PR comments
Move to hashcat

* Address one more small comment

* Will this break something? Who knows

* Move to pip impacket and hashcat for Jenkins

* Use OS default temp dir for Jenkins analysis

* Workaround for hashcat on Docker (I think)

* Trying a fix

* Oops

* Move subprocess to execute

* Move task back to raw disks

* Update dependencies to work
Don't process Windows accounts unless all files are extracted

* PR fixes

* Remove extraneous argument

* Some extra passwords to be on the safe side

* Return if error

* Increase test timeout ??

+++ turbinia/lib/utils.py
@@ -37,7 +37,7 @@ def _image_export(command, output_dir, timeout=DEFAULT_TIMEOUT):
     output_dir: Path to directory to store the the extracted files.
 
   Returns:
-    dict: file names and paths to extracted files.
+    list: paths to extracted files.
 
   Raises:
     TurbiniaException: If an error occurs when running image_export.
@@ -72,7 +72,7 @@ def extract_artifacts(artifact_names, disk_path, output_dir):
     output_dir: Path to directory to store the the extracted files.
 
   Returns:
-    dict: file names and paths to extracted files.
+    list: paths to extracted files.
 
   Raises:
     TurbiniaException: If an error occurs when running image_export.
@@ -97,7 +97,7 @@ def extract_files(file_name, disk_path, output_dir):
     output_dir: Path to directory to store the the extracted files.
 
   Returns:
-    dict: file names and paths to extracted files.
+    list: paths to extracted files.
 
   Raises:
     TurbiniaException: If an error occurs when running image_export.
@@ -128,12 +128,15 @@ def get_exe_path(filename):
   return binary
 
 
-def bruteforce_password_hashes(password_hashes, timeout=300):
-  """Bruteforce password hashes using John the Ripper.
+def bruteforce_password_hashes(
+    password_hashes, tmp_dir, timeout=300, extra_args=''):
+  """Bruteforce password hashes using Hashcat.
 
   Args:
     password_hashes (list): Password hashes as strings.
+    tmp_dir (str): Path to use as a temporary directory
     timeout (int): Number of seconds to run for before terminating the process.
+    extra_args (str): Any extra arguments to be passed to Hashcat.
 
   Returns:
     list: of tuples with hashes and plain text passwords.
@@ -146,7 +149,22 @@ def bruteforce_password_hashes(password_hashes, timeout=300):
     password_hashes_file_path = fh.name
     fh.write('\n'.join(password_hashes))
 
-  cmd = ['john', password_hashes_file_path]
+  pot_file = os.path.join((tmp_dir or tempfile.gettempdir()), 'hashcat.pot')
+  password_list_file_path = os.path.expanduser('~/password.lst')
+
+  # Fallback
+  if not os.path.isfile(password_list_file_path):
+    password_list_file_path = '/usr/share/john/password.lst'
+
+  # Bail
+  if not os.path.isfile(password_list_file_path):
+    raise TurbiniaException('No password list available')
+
+  cmd = ['hashcat', '--force', '-a', '0']
+  if extra_args:
+    cmd = cmd + extra_args.split(' ')
+  cmd = cmd + ['--potfile-path={}'.format(pot_file)]
+  cmd = cmd + [password_hashes_file_path, password_list_file_path]
 
   with open(os.devnull, 'w') as devnull:
     try:
@@ -157,18 +175,18 @@ def bruteforce_password_hashes(password_hashes, timeout=300):
       # Cancel the timer if the process is done before the timer.
       if timer.is_alive():
         timer.cancel()
-    except OSError:
-      raise TurbiniaException('john the ripper failed.')
+    except OSError as e:
+      raise TurbiniaException('hashcat failed: {0}'.format(str(e)))
 
   result = []
-  # Default location of the result file, no way to change it.
-  pot_file = os.path.expanduser('~/.john/john.pot')
 
   if os.path.isfile(pot_file):
     with open(pot_file, 'r') as fh:
-      for line in fh.readlines():
+      for line in fh:
         password_hash, plaintext = line.rsplit(':', 1)
-        result.append((password_hash, plaintext.rstrip()))
+        plaintext = plaintext.rstrip()
+        if plaintext:
+          result.append((password_hash, plaintext))
     os.remove(pot_file)
 
   return result
